+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Fix race condition when TTFD and app launch profiling are both enabled (#5300)
+
 ## 8.52.0-beta
 
 ### Features